Fixing int/size_t comparison warning.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@3265 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Connector/ConnectCCs.h
@@ -267,7 +267,7 @@ void ConnectCCs<CFG,WEIGHT>::Connect( Roadmap<CFG, WEIGHT>* _rm, StatClass& _sta
   if(ccs1.size() <= 1) return;
 
   VIDIT itr1,itr2;
-  for(int i = 0; i < ccs1.size(); ++i){
+  for(size_t i = 0; i < ccs1.size(); ++i){
     ccid.push_back(ccs1[i].second);
   }
